mv test file to unexport metrics, add metric details in middleware doc
