+++ thriftbp/server_middlewares_test.go
@@ -4,17 +4,12 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
-	"strconv"
-	"strings"
 	"testing"
 	"time"
 
 	"github.com/apache/thrift/lib/go/thrift"
-	"github.com/prometheus/client_golang/prometheus/testutil"
 
 	"github.com/reddit/baseplate.go/ecinterface"
-	"github.com/reddit/baseplate.go/internal/gen-go/reddit/baseplate"
 	"github.com/reddit/baseplate.go/mqsend"
 	"github.com/reddit/baseplate.go/thriftbp"
 	"github.com/reddit/baseplate.go/thriftbp/thrifttest"
@@ -336,86 +331,4 @@ func TestPanicMiddleware(t *testing.T) {
 			t.Error("expected an error, got nil")
 		}
 	})
-}
-
-func TestPrometheusMetricsMiddleware(t *testing.T) {
-	testCases := []struct {
-		name          string
-		wantErr       thrift.TException
-		wantOK        bool
-		baseplateCode string
-	}{
-		{"bp error", thrift.NewTProtocolExceptionWithType(thrift.PROTOCOL_ERROR, thriftbp.WrapBaseplateError(errors.New("test"))), false, ""},
-		{"application error", thrift.NewTApplicationException(thrift.UNKNOWN_METHOD, "unknown err msg"), false, ""},
-		{"compile error", baseplate.NewError(), false, "0"},
-		{"success", nil, true, ""},
-	}
-
-	const (
-		serviceName = "testservice"
-		method      = "testmethod"
-	)
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			thriftbp.LatencyDistribution.Reset()
-			thriftbp.RPCStatusCounter.Reset()
-			thriftbp.ActiveRequests.Reset()
-
-			next := thrift.WrappedTProcessorFunction{
-				Wrapped: func(ctx context.Context, seqId int32, in, out thrift.TProtocol) (bool, thrift.TException) {
-					return tt.wantOK, tt.wantErr
-				},
-			}
-			promMiddleware := thriftbp.PrometheusMetricMiddleware(serviceName)
-			wrapped := promMiddleware(method, next)
-			gotOK, gotErr := wrapped.Process(context.Background(), 1, nil, nil)
-
-			if gotOK != tt.wantOK {
-				t.Errorf("wanted %v, got %v", tt.wantOK, gotOK)
-			}
-			if gotErr != tt.wantErr {
-				t.Errorf("wanted %v, got %v", tt.wantErr, gotErr)
-			}
-
-			var baseplateCodeStatus string
-			var exceptionType string
-			if gotErr != nil {
-				exceptionType = strings.TrimPrefix(fmt.Sprintf("%T", gotErr), "*")
-			}
-			success := strconv.FormatBool(gotErr == nil)
-			thriftLabelValues := []string{
-				serviceName,
-				method,
-				success,
-				exceptionType,
-				baseplateCodeStatus,
-				tt.baseplateCode,
-			}
-			latencyMetricCount := testutil.CollectAndCount(thriftbp.LatencyDistribution)
-			if latencyMetricCount != 1 {
-				t.Errorf("wanted %v, got %v", 1, latencyMetricCount)
-			}
-			rpcMetricValue := testutil.ToFloat64(thriftbp.RPCStatusCounter.WithLabelValues(thriftLabelValues...))
-			if rpcMetricValue != 1 {
-				t.Errorf("wanted %v, got %v", 0, rpcMetricValue)
-			}
-			rpcMetricCount := testutil.CollectAndCount(thriftbp.RPCStatusCounter)
-			if rpcMetricCount != 1 {
-				t.Errorf("wanted %v, got %v", 1, rpcMetricCount)
-			}
-
-			activeLabelValues := []string{
-				serviceName,
-				method,
-			}
-			requestsMetricValue := testutil.ToFloat64(thriftbp.ActiveRequests.WithLabelValues(activeLabelValues...))
-			if requestsMetricValue != 0 {
-				t.Errorf("wanted %v, got %v", 0, requestsMetricValue)
-			}
-			requestsMetricCount := testutil.CollectAndCount(thriftbp.RPCStatusCounter)
-			if requestsMetricCount != 1 {
-				t.Errorf("wanted %v, got %v", 1, requestsMetricCount)
-			}
-		})
-	}
 }
